Move Desktop- and MobileComponents to own folders in /src/components
