+++ src/components/DesktopComponent/index.js
@@ -1,0 +1,3 @@
+import DesktopComponent from './DesktopComponent';
+
+export default DesktopComponent;
