+++ src/components/MobileComponent/index.js
@@ -1,0 +1,3 @@
+import MobileComponent from './MobileComponent';
+
+export default MobileComponent;
