+++ src/views/HomeView/HomeView.js
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Map } from '@material-ui/icons';
 import SearchBar from '../../components/SearchBar';
-import { MobileComponent } from '../../layouts/WrapperComponents/WrapperComponents';
 import PaperButton from '../../components/PaperButton';
 import { getIcon } from '../../components/SMIcon';
+import MobileComponent from '../../components/MobileComponent';
 
 class HomeView extends React.Component {
   renderNavigationOptions = () => {
